+++ ledger/Cargo.toml
@@ -28,6 +28,7 @@ frozen-abi = [
 agave-unstable-api = []
 
 [dependencies]
+agave-cluster-transport = { workspace = true }
 agave-feature-set = { workspace = true }
 agave-reserved-account-keys = { workspace = true }
 agave-snapshots = { workspace = true }
@@ -98,7 +99,6 @@ solana-signer = { workspace = true }
 solana-stake-interface = { workspace = true, features = ["wincode"] }
 solana-storage-bigtable = { workspace = true }
 solana-storage-proto = { workspace = true }
-solana-streamer = { workspace = true }
 solana-svm = { workspace = true, features = ["metrics"] }
 solana-svm-timings = { workspace = true }
 solana-svm-transaction = { workspace = true }
