+++ ledger/src/blockstore/tests.rs
@@ -8,6 +8,7 @@ use {
             merkle_tree::{SIZE_OF_MERKLE_PROOF_ENTRY, get_proof_size, verify_merkle_proof},
         },
     },
+    agave_cluster_transport::evicting_sender::EvictingSender,
     assert_matches::assert_matches,
     rand::{rng, seq::SliceRandom},
     solana_entry::entry::next_entry_mut,
@@ -20,7 +21,6 @@ use {
     solana_sha256_hasher::hash,
     solana_shred_version::version_from_hash,
     solana_signature::Signature,
-    solana_streamer::evicting_sender::EvictingSender,
     solana_transaction::Transaction,
     solana_transaction_context::transaction::TransactionReturnData,
     solana_transaction_error::TransactionError,
