+++ streamer/src/lib.rs
@@ -1,14 +1,16 @@
 #![cfg(feature = "agave-unstable-api")]
 #![allow(clippy::arithmetic_side_effects)]
-pub mod evicting_sender;
-pub mod msghdr;
 pub mod nonblocking;
-pub mod packet;
 pub mod quic;
 pub mod quic_socket;
-pub mod recvmmsg;
-pub mod sendmmsg;
-pub mod streamer;
+
+pub use agave_cluster_transport::{evicting_sender, packet, recvmmsg, sendmmsg};
+
+pub mod streamer {
+    pub use agave_cluster_transport::{
+        packet::filter_packets_by_socket_addr_space, receiver::*, staked_nodes::StakedNodes,
+    };
+}
 
 #[macro_use]
 extern crate log;
