+++ streamer/src/nonblocking/quic.rs
@@ -6,8 +6,8 @@ use {
         },
         quic::{QuicServerError, QuicStreamerConfig, StreamerStats, configure_server},
         quic_socket::{QuicSocket, QuicXdpSocketParts, QuicXdpTxSocket},
-        streamer::StakedNodes,
     },
+    agave_cluster_transport::staked_nodes::StakedNodes,
     bytes::{BufMut, Bytes, BytesMut},
     crossbeam_channel::{Sender, TrySendError},
     futures::{Future, StreamExt as _, stream::FuturesUnordered},
