+++ streamer/src/nonblocking/simple_qos.rs
@@ -13,8 +13,8 @@ use {
             DEFAULT_MAX_QUIC_CONNECTIONS_PER_STAKED_PEER, DEFAULT_MAX_STAKED_CONNECTIONS,
             DEFAULT_MAX_STREAMS_PER_MS, StreamerStats,
         },
-        streamer::StakedNodes,
     },
+    agave_cluster_transport::staked_nodes::StakedNodes,
     quinn::{Connection, VarInt},
     solana_net_utils::{banlist::Banlist, token_bucket::TokenBucket},
     solana_pubkey::Pubkey,
@@ -435,7 +435,6 @@ mod tests {
                 testing_utilities::get_client_config,
             },
             quic::{QuicStreamerConfig, StreamerStats, configure_server},
-            streamer::StakedNodes,
         },
         quinn::Endpoint,
         solana_keypair::{Keypair, Signer},
