+++ udp-client/Cargo.toml
@@ -19,17 +19,16 @@ agave-unstable-api = []
 dev-context-only-utils = []
 
 [dependencies]
+agave-cluster-transport = { workspace = true }
 async-trait = { workspace = true }
 solana-connection-cache = { workspace = true }
 solana-keypair = { workspace = true }
 solana-net-utils = { workspace = true }
-solana-streamer = { workspace = true }
 solana-transaction-error = { workspace = true }
 
 [dev-dependencies]
 solana-net-utils = { path = "../net-utils", features = ["agave-unstable-api", "dev-context-only-utils"] }
 solana-packet = { workspace = true }
-solana-streamer = { path = "../streamer", features = ["agave-unstable-api", "dev-context-only-utils"] }
 
 [lints]
 workspace = true
